Added read tests

+++ src/main/scala/prog/Point.scala
@@ -1,6 +1,6 @@
 package prog
 
-case class Point(x: Float, y: Float, r: Int) {
+case class Point(x: Double, y: Double, r: Int) {
   def check(): Boolean = {
     if (x >= 0) {
       if (y >= 0) { if (x * x + y * y <= r * r) return true }
